+++ pages/[id]/index.js
@@ -3,9 +3,9 @@ import SnippetDetails from "@/components/SnippetDetails";
 import { useRouter } from "next/router";
 import { mutate } from "swr";
 
-function SnippetDetailsPage({ editState }) {
+function SnippetDetailsPage() {
   const router = useRouter();
-  // const { isReady } = router;
+
   const { id } = router.query;
 
   async function handleDelete() {
@@ -17,7 +17,6 @@ function SnippetDetailsPage({ editState }) {
   return (
     <>
       <BackLink url={"/"} />
-      {editState && <p>Editted successfully</p>}
       <SnippetDetails onDelete={handleDelete} />
     </>
   );
